# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Optivote
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/password/reset:
    post:
      summary: 'Reset password.'
      operationId: resetPassword
      description: 'This endpoint reset user password.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Mot de passe réinitialisé avec succès'
                  body: null
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Mot de passe réinitialisé avec succès'
                  body:
                    type: string
                    example: null
        400:
          description: ''
          content:
            application/json:
              schema:
                oneOf:
                  -
                    description: ''
                    type: object
                    example:
                      success: false
                      message: 'OTP invalide ou expiré.'
                      body: null
                    properties:
                      success:
                        type: boolean
                        example: false
                      message:
                        type: string
                        example: 'OTP invalide ou expiré.'
                      body:
                        type: string
                        example: null
                  -
                    description: ''
                    type: object
                    example:
                      success: false
                      message: 'OTP expiré.'
                      body: null
                    properties:
                      success:
                        type: boolean
                        example: false
                      message:
                        type: string
                        example: 'OTP expiré.'
                      body:
                        type: string
                        example: null
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'The email address of the user.'
                  example: user@example.com
                  nullable: false
                otp:
                  type: numeric
                  description: ''
                  example: deserunt
                  nullable: false
                password:
                  type: string
                  description: 'minimum six characters'
                  example: 's@D|\*:}6SSu,xTC@'
                  nullable: false
                password_confirmation:
                  type: string
                  description: 'equal to password'
                  example: consequuntur
                  nullable: false
              required:
                - email
                - otp
                - password
                - password_confirmation
      security: []
...